requirements.txt
streamlit==1.44.1
streamlit-option-menu==0.4.0
streamlit-slickgrid==0.1.11
pillow==11.2.1